JeeLib support for energy savings
Add JeeLib Library support, and minor changes in the code to save as
much energy as possible to run on battery.

Change hardware pins to adapt it to your RF Tx and DHT pinout.

Set it to debug mode to enable serial monitor and LED blink.
In production mode it is better to remove the power status LED from the
Arduino PCB.